+++ apps/database/migrations/000016_add_audit_log.up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE audit_log (
   id bigserial PRIMARY KEY,
-  user_id integer NOT NULL REFERENCES users ON DELETE CASCADE,
+  user_id bigint NOT NULL REFERENCES users ON DELETE CASCADE,
   action varchar(50) NOT NULL,
   created timestamptz NOT NULL DEFAULT now()
 );
